Add "smart resize" function to sort out the resamplers and downscalers for you
Application code can now call the SetSmartResize method for the output
branches which will ask the library to sort out the arangement of
resamples and downscalers for you, in a (hopefully) optimal manner.

If smart resize is not used, the previous behaviour, where you had to
figure everything out for yourself, is maintained.