Tiling reliability improvements
We allow right hand edge tiles to be smaller than 16 pixels in the
output branches (after the split and crop). There are always enough
pixels available there for context so that this works.

This makes it unnecessary to pull penultimate tiles back from the RH
edge when we were "close but not quite there". In practice, alignment
constraints meant that it didn't work anyway.