backend: Export/import the Backend config as JSON
Add a new helper BackEnd::GetJsonConfig() to return a JSON string of
the current Backend register configuration and tile data. The parameters
are the actual register values to be programmed into the hardware, and
currently all dumped as unsigned.

Add a new helper BackEnd::SetJsonConfig() to read a JSON string and
import the Backend register configuration. Tile data does not get
imported, and will need a BackEnd::Prepare() call to re-generate the
tile data.

Signed-off-by: Naushir Patuck <naush@raspberrypi.com>