backend: Allow local builds to use the src backend_default_config.json
If an application links to a local install of libpisp.so, the backend
will fail to find backend_default_config.json in the data install path.

Fix this by testing if libpisp is installed, and if not, use the build
path to find backend_default_config.json, which is now copied to the
specific location by the build scripts.

This functionality needs linkage with libdl as well as a custom rpath
that gets stripped with a meson install step.

Signed-off-by: Naushir Patuck <naush@raspberrypi.com>