backend: frontend: Make BackEnd and FrontEnd standard layout types
To allow the BackEnd and FrontEnd objects to be safely shared between
processes, it must be a standard layout type.  Add a static_assert to
ensure this is the case.

Remove the const reference PiSPVariant members from both BackEnd and
FrontEnd and replace with a const PiSPVariant member.

Remove all BackEnd default config std::map types.  These are replaced by
std::vector<std::pair<.,.>> types which are standard layout types.

Signed-off-by: Naushir Patuck <naush@raspberrypi.com>