tiling: Track the branch completion status via the output stages
Tiles may produce output on only a single branch when multiple branches
are active. Track the completion using a new BranchComplete() helper,
which traverses the pipeline till the output stage and returns a flag
if the stage has completed all its output.

In the split stage, do not inject tiles into a branch if the output is
complete.

Signed-off-by: Naushir Patuck <naush@raspberrypi.com>