libpisp: Workaround -Wno-error-address-of-packed-member
The mainline version of pisp_be_config.h and pisp_common.h define
all types that are to be written directly to HW registers as
packed, to avoid the compiler inserting holes to force any alignment
of the types' members.

However, this triggers a compiler error when the types members are

error: taking address of packed member of ‘pisp_image_format_config’
may result in an unaligned pointer value
[-Werror=address-of-packed-member]

See https://gcc.gnu.org/bugzilla/show_bug.cgi?id=36566#c6 for a more
detailed explanation about the compiler error is actually correct.

Albeit libpisp is only going to be compiled on architectures that allow
byte-alligned access (ARMv8) the compiler flags accesses by pointer to
packed struct members as errors.

Work this around by using a temporary variable to store the packed
structure member value and pass it by pointer. As the functions the
variables are passed to might modify the variable values, re-assign them
after the function call to the packed structure members.

Signed-off-by: Jacopo Mondi <jacopo.mondi@ideasonboard.com>