tiling: More robust handling of multiple crop windows
Remove the started flag from the crop stage and rely on the output
interval length instead. Add a BranchInactive() helper to the base stage
class.

When testing for the above, ensure we handle the case where the output
interval end is > min tile size, but the length is not because of a
possible negative offset.

In BasicStage::CopyOut(), if a stage is inactive or complete, reset all
intervals before copying out to the tile structures.

Signed-off-by: Naushir Patuck <naush@raspberrypi.com>