utils: Preserve API by overloading libpisp::compute_optimal_stride
The signature for compute_optimal_stride has changed, and it is called by libcamera
extrenally. While adding a default parameter makes it backwards compatible, using
libcamera.so to build rpicam-apps with -Wl, --no-undefined will cause undefined symbols.

This change preserves the API by overloading the function to have both the previous and
new signature. This will emit both symbols in the shared object. Only
compute_optimal_stride in utils.hpp needs this change as this is the only function being
used in libcamera as an API from utils.