examples: convert: Add support for RGBX output from 2712C1
By doubling the width, changing the output format to "UYVY" and
applying a matrix to swap the first two colour components, we
can output RGBX. The gotcha is that when downscaling, this can
result in an oversharp/aliasy image.

No support for RGBX input (though that would theoretically be
possible at unity scale, using a customized Resample kernel).

Signed-off-by: Nick Hollinghurst <nick.hollinghurst@raspberrypi.com>